Create POC_Modules_Distribution (examples)
Need lib: gson-2.13.1.jar
module-info.properties
     exports
     requires
     provides
     requires transite
     opens
API gson -  convert Object -> Json / Json -> Object (Serializer and Deserializer)
command line -> jdeps: show which modules, packages  or classes the code depends on
command line -> jlink: tool that allows you to create a customized JVM containing only the modules needed to run application.
jlink laucher: It is an automatically created binary that runs the application, as if it were an exclusive command for it.